Move ledger package back to horizon internals
