+++ services/horizon/internal/actions_root.go
@@ -2,7 +2,7 @@ package horizon
 
 import (
 	"github.com/stellar/go/protocols/horizon"
-	"github.com/stellar/go/protocols/horizon/ledger"
+	"github.com/stellar/go/services/horizon/internal/ledger"
 	"github.com/stellar/go/services/horizon/internal/resourceadapter"
 	"github.com/stellar/go/support/render/hal"
 )
