+++ services/horizon/internal/app.go
@@ -12,10 +12,10 @@ import (
 	metrics "github.com/rcrowley/go-metrics"
 	"github.com/stellar/go/build"
 	"github.com/stellar/go/clients/stellarcore"
-	"github.com/stellar/go/protocols/horizon/ledger"
 	"github.com/stellar/go/services/horizon/internal/db2/core"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
 	"github.com/stellar/go/services/horizon/internal/ingest"
+	"github.com/stellar/go/services/horizon/internal/ledger"
 	"github.com/stellar/go/services/horizon/internal/log"
 	"github.com/stellar/go/services/horizon/internal/paths"
 	"github.com/stellar/go/services/horizon/internal/reap"
